move app-level user auth token logic into its own file
