+++ server/remote/msgraph/get_app_level_token.go
@@ -1,0 +1,40 @@
+// Copyright (c) 2019-present Mattermost, Inc. All Rights Reserved.
+// See License for license information.
+
+package msgraph
+
+import (
+	"net/http"
+	"net/url"
+)
+
+type AuthResponse struct {
+	TokenType   string `json:"token_type"`
+	ExpiresIn   int    `json:"expires_in"`
+	AccessToken string `json:"access_token"`
+}
+
+func (c *client) getAppLevelToken() (string, error) {
+	params := map[string]string{
+		"client_id":     c.conf.OAuth2ClientID,
+		"scope":         "https://graph.microsoft.com/.default",
+		"client_secret": c.conf.OAuth2ClientSecret,
+		"grant_type":    "client_credentials",
+	}
+
+	u := "https://login.microsoftonline.com/" + c.conf.OAuth2Authority + "/oauth2/v2.0/token"
+	res := AuthResponse{}
+
+	data := url.Values{}
+	data.Set("client_id", params["client_id"])
+	data.Set("scope", params["scope"])
+	data.Set("client_secret", params["client_secret"])
+	data.Set("grant_type", params["grant_type"])
+
+	_, err := c.Call(http.MethodPost, u, data, &res)
+	if err != nil {
+		return "", err
+	}
+
+	return res.AccessToken, nil
+}
